+++ redis/src/lib.rs
@@ -21,14 +21,14 @@
 //!         let mut conn = pool.get().await.unwrap();
 //!         cmd("SET")
 //!             .arg(&["deadpool/test_key", "42"])
-//!             .execute(&mut conn)
+//!             .execute_async(&mut conn)
 //!             .await.unwrap();
 //!     }
 //!     {
 //!         let mut conn = pool.get().await.unwrap();
 //!         let value: String = cmd("GET")
 //!             .arg(&["deadpool/test_key"])
-//!             .query(&mut conn)
+//!             .query_async(&mut conn)
 //!             .await.unwrap();
 //!         assert_eq!(value, "42".to_string());
 //!     }
@@ -36,8 +36,9 @@
 //! ```
 #![warn(missing_docs)]
 
+use std::ops::{Deref, DerefMut};
+
 use async_trait::async_trait;
-use futures::compat::Future01CompatExt;
 use redis::{
     aio::Connection as RedisConnection, Client, IntoConnectionInfo, RedisError, RedisResult,
 };
@@ -49,7 +50,6 @@ pub type Pool = deadpool::managed::Pool<ConnectionWrapper, RedisError>;
 pub type PoolError = deadpool::managed::PoolError<RedisError>;
 
 type RecycleResult = deadpool::managed::RecycleResult<RedisError>;
-type RecycleError = deadpool::managed::RecycleError<RedisError>;
 
 mod cmd_wrapper;
 pub use cmd_wrapper::{cmd, Cmd};
@@ -63,23 +63,19 @@ pub use pipeline_wrapper::{pipe, Pipeline};
 /// This wrapper makes it possible to replace the internal connection after
 /// executing a query.
 pub struct ConnectionWrapper {
-    conn: Option<RedisConnection>,
+    conn: RedisConnection,
 }
 
-impl ConnectionWrapper {
-    fn _take_conn(&mut self) -> RedisResult<RedisConnection> {
-        if let Some(conn) = self.conn.take() {
-            Ok(conn)
-        } else {
-            Err(redis::RedisError::from((
-                redis::ErrorKind::IoError,
-                "deadpool.redis: Connection to server lost due to previous query",
-            )))
-        }
+impl Deref for ConnectionWrapper {
+    type Target = RedisConnection;
+    fn deref(&self) -> &RedisConnection {
+        &self.conn
     }
+}
 
-    fn _replace_conn(&mut self, conn: RedisConnection) {
-        self.conn = Some(conn)
+impl DerefMut for ConnectionWrapper {
+    fn deref_mut(&mut self) -> &mut RedisConnection {
+        &mut self.conn
     }
 }
 
@@ -100,20 +96,14 @@ impl Manager {
 #[async_trait]
 impl deadpool::managed::Manager<ConnectionWrapper, RedisError> for Manager {
     async fn create(&self) -> Result<ConnectionWrapper, RedisError> {
-        let conn = self.client.get_async_connection().compat().await?;
-        Ok(ConnectionWrapper { conn: Some(conn) })
+        let conn = self.client.get_async_connection().await?;
+        Ok(ConnectionWrapper { conn })
     }
 
     async fn recycle(&self, conn: &mut ConnectionWrapper) -> RecycleResult {
-        if conn.conn.is_some() {
-            match cmd("PING").execute(conn).await {
-                Ok(_) => Ok(()),
-                Err(e) => Err(e.into()),
-            }
-        } else {
-            Err(RecycleError::Message(
-                "deadpool.redis: Connection could not be recycled".to_string(),
-            ))
+        match cmd("PING").execute_async(conn).await {
+            Ok(_) => Ok(()),
+            Err(e) => Err(e.into()),
         }
     }
 }